+++ Editor/Drawers/PropRangeDrawer/PropRangeAttributeDrawerUIToolkit.cs
@@ -130,82 +130,83 @@ namespace SaintsField.Editor.Drawers.PropRangeDrawer
             AdaptAttribute adaptAttribute = allAttributes.OfType<AdaptAttribute>().FirstOrDefault();
 
             PropRangeField propRangeField = container.Q<PropRangeField>(name: NamePropRange(property));
-            propRangeField.labelElement.AddManipulator(new ContextualMenuManipulator(evt =>
-            {
-                evt.menu.AppendAction("Copy Property Path", _ => EditorGUIUtility.systemCopyBuffer = property.propertyPath);
-#if SAINTSFIELD_NEWTONSOFT_JSON
-                evt.menu.AppendSeparator();
-                evt.menu.AppendAction("Copy", _ =>
-                {
-                    string content;
-                    // ReSharper disable once ConvertIfStatementToConditionalTernaryExpression
-                    if (property.propertyType == SerializedPropertyType.Integer)
-                    {
-                        content = ClipboardUtils.CopyGenericType(property.intValue);
-                    }
-                    else
-                    {
-                        content = ClipboardUtils.CopyGenericType(property.floatValue);
-                    }
-                    EditorGUIUtility.systemCopyBuffer = content;
-                });
-
-                JToken jtoken;
-                try
-                {
-                    jtoken = JToken.Parse(EditorGUIUtility.systemCopyBuffer);
-                }
-                catch (Exception)
-                {
-                    evt.menu.AppendAction("Paste", _ => { }, DropdownMenuAction.Status.Disabled);
-                    return;
-                }
-
-                int intResult = 0;
-                float floatResult = 0;
-
-                // ReSharper disable once SwitchStatementHandlesSomeKnownEnumValuesWithDefault
-                switch (jtoken.Type)
-                {
-                    case JTokenType.Integer:
-                        intResult = jtoken.Value<int>();
-                        break;
-                    case JTokenType.Float:
-                        floatResult = jtoken.Value<float>();
-                        break;
-                    default:
-                        evt.menu.AppendAction("Paste", _ => { }, DropdownMenuAction.Status.Disabled);
-                        return;
-                }
-
-                evt.menu.AppendAction("Paste", _ =>
-                {
-                    if (property.propertyType == SerializedPropertyType.Integer)
-                    {
-#if SAINTSFIELD_DEBUG
-                        Debug.Log($"Pasted: {intResult}");
-#endif
-                        property.intValue = intResult;
-                    }
-                    else
-                    {
-#if SAINTSFIELD_DEBUG
-                        Debug.Log($"Pasted: {floatResult}");
-#endif
-                        property.floatValue = floatResult;
-                    }
-                    property.serializedObject.ApplyModifiedProperties();
-                    if (property.propertyType == SerializedPropertyType.Integer)
-                    {
-                        onValueChangedCallback.Invoke(intResult);
-                    }
-                    else
-                    {
-                        onValueChangedCallback.Invoke(floatResult);
-                    }
-                });
-#endif
-            }));
+            UIToolkitUtils.AddContextualMenuManipulator(propRangeField.labelElement, property, () => Util.PropertyChangedCallback(property, info, onValueChangedCallback));
+//             propRangeField.labelElement.AddManipulator(new ContextualMenuManipulator(evt =>
+//             {
+//                 evt.menu.AppendAction("Copy Property Path", _ => EditorGUIUtility.systemCopyBuffer = property.propertyPath);
+// #if SAINTSFIELD_NEWTONSOFT_JSON
+//                 evt.menu.AppendSeparator();
+//                 evt.menu.AppendAction("Copy", _ =>
+//                 {
+//                     string content;
+//                     // ReSharper disable once ConvertIfStatementToConditionalTernaryExpression
+//                     if (property.propertyType == SerializedPropertyType.Integer)
+//                     {
+//                         content = ClipboardUtils.CopyGenericType(property.intValue);
+//                     }
+//                     else
+//                     {
+//                         content = ClipboardUtils.CopyGenericType(property.floatValue);
+//                     }
+//                     EditorGUIUtility.systemCopyBuffer = content;
+//                 });
+//
+//                 JToken jtoken;
+//                 try
+//                 {
+//                     jtoken = JToken.Parse(EditorGUIUtility.systemCopyBuffer);
+//                 }
+//                 catch (Exception)
+//                 {
+//                     evt.menu.AppendAction("Paste", _ => { }, DropdownMenuAction.Status.Disabled);
+//                     return;
+//                 }
+//
+//                 int intResult = 0;
+//                 float floatResult = 0;
+//
+//                 // ReSharper disable once SwitchStatementHandlesSomeKnownEnumValuesWithDefault
+//                 switch (jtoken.Type)
+//                 {
+//                     case JTokenType.Integer:
+//                         intResult = jtoken.Value<int>();
+//                         break;
+//                     case JTokenType.Float:
+//                         floatResult = jtoken.Value<float>();
+//                         break;
+//                     default:
+//                         evt.menu.AppendAction("Paste", _ => { }, DropdownMenuAction.Status.Disabled);
+//                         return;
+//                 }
+//
+//                 evt.menu.AppendAction("Paste", _ =>
+//                 {
+//                     if (property.propertyType == SerializedPropertyType.Integer)
+//                     {
+// #if SAINTSFIELD_DEBUG
+//                         Debug.Log($"Pasted: {intResult}");
+// #endif
+//                         property.intValue = intResult;
+//                     }
+//                     else
+//                     {
+// #if SAINTSFIELD_DEBUG
+//                         Debug.Log($"Pasted: {floatResult}");
+// #endif
+//                         property.floatValue = floatResult;
+//                     }
+//                     property.serializedObject.ApplyModifiedProperties();
+//                     if (property.propertyType == SerializedPropertyType.Integer)
+//                     {
+//                         onValueChangedCallback.Invoke(intResult);
+//                     }
+//                     else
+//                     {
+//                         onValueChangedCallback.Invoke(floatResult);
+//                     }
+//                 });
+// #endif
+//             }));
 
             Slider slider = propRangeField.Q<Slider>(NameSlider(property));